requirements.txt
PyYAML
fastapi
uvicorn
paho-mqtt